Move MESSAGE_UNKNOWN_COMMAND to AddressBookParser
AddressBookParser is the only user of this message.

Furthermore, AddressBookParser is the parser that implements the concept
of "individual commands" in the command line. It is the parser that
looks at the first word of the command line, and then dispatches the
rest of the command line to the associated command parser based on that
word. As such, AddressBookParser is very likely the only class that
would ever use the message "Unknown command", now and in the future.

As such, let's move MESSAGE_UNKNOWN_COMMAND to AddressBookParser.
