+++ src/main/java/seedu/address/logic/parser/AddressBookParser.java
@@ -1,7 +1,6 @@
 package seedu.address.logic.parser;
 
 import static seedu.address.commons.core.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
-import static seedu.address.commons.core.Messages.MESSAGE_UNKNOWN_COMMAND;
 
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
@@ -25,6 +24,8 @@ import seedu.address.logic.parser.exceptions.ParseException;
  * Parses user input.
  */
 public class AddressBookParser {
+
+    public static final String MESSAGE_UNKNOWN_COMMAND = "Unknown command";
 
     /**
      * Used for initial separation of command word and args.
